Update installation prompt for desktop utilities
Refactor desktop utilities installation prompt to check for interactive shell.

+++ setup_env.sh
@@ -211,19 +211,26 @@ fi
 echo
 echo "All done."
 
-while true; do
-    read -r -p 2 -p "Install desktop utilities ? [Y/n]" yn
-    case $yn in
-        [Yy]* ) 
-            sudo apt install libfuse2
-            wget -O - https://raw.githubusercontent.com/laurent22/joplin/dev/Joplin_install_and_update.sh | bash
-            curl -fsS https://dl.brave.com/install.sh | sh
-            curl -f https://zed.dev/install.sh | sh; break;;
-        [Nn]* ) exit;;
-        * ) echo "Please answer yes or no.";;
-    esac
-done
-
+if [[ $- == *i* ]]
+then
+    while true; do
+        read -r -p 2 -p "Install desktop utilities ? [Y/n]" yn
+        case $yn in
+            [Yy]* ) 
+                sudo apt install libfuse2
+                wget -O - https://raw.githubusercontent.com/laurent22/joplin/dev/Joplin_install_and_update.sh | bash
+                curl -fsS https://dl.brave.com/install.sh | sh
+                curl -f https://zed.dev/install.sh | sh; break;;
+            [Nn]* ) exit;;
+            * ) echo "Please answer yes or no.";;
+        esac
+    done
+    else
+        echo "Install desktop utilities manually"
+        echo "wget -O - https://raw.githubusercontent.com/laurent22/joplin/dev/Joplin_install_and_update.sh | bash"
+        echo "curl -fsS https://dl.brave.com/install.sh | sh"
+        echo "curl -f https://zed.dev/install.sh | sh"
+fi    
     
 
 echo "Backups are stored in: $BACKUP_DIR"
